Update readme.md
<href=http//www.domanregistry.de><IMG SRC=http://www.domainregistry.de/dev-domains.png></a>

The Dev-Domains are open domains: anyone can register Dev-Domains. The Dev-Domain is meant for developers.

The ICANN process of introducing new Top Level Domains continues. The registry of the Dev-Domains has set the 16th of January 2019 as the start of the Sunrise Period.

Developments have now their own more secure home on the web: the Dev-Domain, the most relevant TLD for website and software developers.

The relationship between a better ranking in search engines and new top-level domains has already been proven in a study of Searchmetrics for .berlin domains. Websites with .berlin domains often rank better with regional searches in Google than sites with .de domains and .com domains.

»In 42% of searches, .berlin domains rank better locally.«

Another study by Total Websites in Houston shows that the results of the Searchmetrics study are in principle applicable to all new TLDs, including e.g. new Dev-Domains and App-Domains. Total Websites notes that Google sees the domain extensions of new TLDs as an important criterion for the evaluation of a domain and, therefore, concludes that:

»It's clear that the new top-level domains improve the ranking in search engines.«



http://www.domainregistry.de/dev-domain.html

http://www.domainregistry.de/app-domain.html

http://www.domainregistry.de/software-domain.html